move json.SafeSet
